grist-org-visualizer-gpt/pyproject.toml
[project]
name = "grist-org-visualizer"
version = "0.1.0"
description = "Generate SDID PI Planning org visualization (draw.io), fragmentation analysis, and synthesis PowerPoint from Grist."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "python-pptx>=0.6.23",
  "requests>=2.31",
  "PyYAML>=6.0",
]

[project.scripts]
grist-orgviz = "src.cli:main"

[tool.ruff]
line-length = 100
